requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
requests
redis
django-cors-headers